+++ R/profit_factor.R
@@ -12,7 +12,8 @@
 #'
 #' @examples
 #'
-#' library(tidyverse)
+#' library(magrittr)
+#' library(dplyr)
 #' data(profit_factor_data)
 #' profit_factor_data %>%
 #'    group_by(X1) %>%
@@ -23,7 +24,7 @@
 #'
 #'
 profit_factor <- function(x){
-  requireNamespace("tidyverse", quietly = TRUE)
+
   # calculate profit factor
   res <- sum(x[x>0])/(0.0001+sum(abs(x[x<0])))