+++ R/write_control_parameters.R
@@ -18,7 +18,8 @@
 #'
 #' \donttest{
 #' #test lasts 15 sec:
-#' library(tidyverse)
+#' library(dplyr)
+#' library(readr)
 #' library(ReinforcementLearning)
 #' library(magrittr)
 #' data(data_trades)
@@ -28,7 +29,8 @@
 #'
 write_control_parameters <- function(x, path_control_files){
 
-  requireNamespace("tidyverse", quietly = TRUE)
+  requireNamespace("dplyr", quietly = TRUE)
+  requireNamespace("readr", quietly = TRUE)
   requireNamespace("ReinforcementLearning", quietly = TRUE)
   requireNamespace("magrittr", quietly = TRUE)