+++ tests/testthat/test-aml_test_model.R
@@ -1,5 +1,6 @@
 library(testthat)
-library(tidyverse)
+library(dplyr)
+library(magrittr)
 library(lazytrade)