+++ tests/testthat/test-import_data.R
@@ -1,5 +1,5 @@
 library(testthat)
-library(tidyverse)
+library(readr)
 library(lubridate)